+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Updated Polars to 0.27.1
+- Added `iter_rows` method to `DataFrame`
 - Added `named` option to `row` and `rows` methods
 - Replaced `include_bounds` option with `closed` for `is_between` method